feat: enhance NDCU document handling with new tables mixin and improved PDF methods

+++ src/dengue/ndcu_doc/NDCUDocPDFMixin.py
@@ -12,7 +12,10 @@ class NDCUDocPDFMixin:
     def pdf_file(self) -> PDFFile:
         return PDFFile(os.path.join(self.dir_data, "original.pdf"))
 
-    def build_pdf(self, temp_pdf_file):
+    def build_pdf(self, temp_pdf_file, force=False):
+        if self.pdf_file.exists:
+            log.debug(f"{self.pdf_file} exists")
+            return
         shutil.copy(
             temp_pdf_file.path,
             self.pdf_file.path,
